+++ backend/models/chat_models.py
@@ -1,0 +1,7 @@
+from pydantic import BaseModel
+from typing import List, Dict, Optional
+
+
+class ChatRequest(BaseModel):
+    question: str = "what is the meaning of life?"
+    history: Optional[List[Dict[str, str]]] = []
